v1_0_7
The user should be able to go in and out of the plane mode. Draw a plane grid where the plane exists otherwise people would not notice that it was still there. The plane disappears when the user goes out of the plane mode and reappears at the new position and orientation when the user goes into the plane mode again.