v1
Got the Leap Motion to work perfectly with the Zed Mini. The virtual hand is aligned with the AR hand feed from Zed Mini. Factors that were modified, IPD of rendering for the Leap Motion camera, values in the advanced device offset mode, and actually sliding the Leap Motion camera physically on top of the Zed Mini until it aligned with the AR feed