1_2_6 Finger foreground subtraction
Took the mesh provided by Leap Motion of the hands and considered that to be the foreground. Now only kept the pixels that matched the skin tone, assigned those pixels the depth of the hands. The rest of the pixels were assigned the depth of the same u, but v as equal to zero (assumed that this will approximate the correct depth of the background environment